experiments/requirements.txt
# 多模态内部威胁检测系统依赖
# Multimodal Internal Threat Detection System Dependencies

# 核心依赖 Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 深度学习 Deep Learning
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0

# 数据处理 Data Processing
dask[dataframe]>=2022.0.0
dask[distributed]>=2022.0.0
pyarrow>=8.0.0

# 可视化 Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
bokeh>=2.4.0

# 配置和日志 Configuration and Logging
pyyaml>=6.0
jsonschema>=4.0.0

# 并行处理 Parallel Processing
joblib>=1.1.0
multiprocessing-logging>=0.3.0

# 网络分析 Network Analysis
networkx>=2.6.0

# 文本处理 Text Processing
nltk>=3.7
spacy==3.4.4

# 工具库 Utilities
tqdm>=4.62.0
pathlib2>=2.3.0
python-dateutil>=2.8.0

# 开发和测试 Development and Testing
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# 可选依赖 Optional Dependencies
# GPU支持 GPU Support (uncomment if needed)
# torch-audio>=0.9.0
# torch-text>=0.10.0

# 高级可视化 Advanced Visualization (uncomment if needed)
# plotly>=5.0.0
# dash>=2.0.0

# 模型解释 Model Interpretation (uncomment if needed)
# shap>=0.40.0
# lime>=0.2.0

# 新增实验框架依赖
wandb>=0.12.0              # 实验跟踪和可视化
optuna>=3.0.0              # 超参数优化
imbalanced-learn>=0.8.0    # 数据不平衡处理 (SMOTE等)
shap>=0.41.0               # 特征重要性分析
plotly>=5.0.0              # 交互式可视化
xgboost>=1.5.0             # XGBoost基线模型 